Fixed: Add validations of ownership of posts and comments

+++ src/services/comments.service.ts
@@ -1,3 +1,5 @@
+import User from "../types/interfaces/user.interface";
+import Post from "../models/post.model";
 import Comment from "../models/comment.model";
 import { CreateCommentDTO } from "../types/dtos/create-comment.dto";
 import { UpdateCommentDTO } from "../types/dtos/update-comment.dto";
@@ -7,25 +9,60 @@ export class CommentsService {
     return await Comment.find();
   }
 
-  async create(comment: CreateCommentDTO): Promise<Comment> {
+  async create(id: string, createCommentDTO: CreateCommentDTO, author: User) {
     try {
-      return await Comment.create(comment);
+      const comment = new Comment({
+        ...createCommentDTO,
+        author,
+      });
+
+      await comment.save();
+
+      const postRelated = await Post.findById(id);
+      postRelated.comments.push(comment);
+
+      return await postRelated.save();
     } catch (error) {
       throw new Error(error);
     }
   }
 
-  async update(id: string, comment: UpdateCommentDTO): Promise<Comment> {
+  async update(
+    id: string,
+    comment: UpdateCommentDTO,
+    author: User
+  ): Promise<Comment> {
     try {
-      return await Comment.findByIdAndUpdate(id, comment, { new: true });
+      if (author._id !== comment.author) {
+        throw new Error("You are not the author of this comment");
+      }
+
+      const updatedComment = await Comment.findByIdAndUpdate(
+        id,
+        {
+          ...comment,
+          author,
+        },
+        { new: true }
+      );
+
+      return updatedComment;
     } catch (error) {
       throw new Error(error);
     }
   }
 
-  async delete(id: string): Promise<Comment> {
+  async delete(id: string, author: User): Promise<Comment> {
     try {
-      return await Comment.findByIdAndDelete(id);
+      const comment = await Comment.findById(id);
+
+      if (author._id !== comment.author) {
+        throw new Error("You are not the author of this comment");
+      }
+
+      await Comment.findByIdAndDelete(id);
+
+      return comment;
     } catch (error) {
       throw new Error(error);
     }
